split podinfo implementation to avoid NRI transitive dependencies
